migrated get_user_input
